requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
dj-database-url==2.2.0
Django==4.2.16
django-allauth==65.0.2
django-bootstrap-v5==1.0.11
django-bootstrap5==24.3
django-debug-toolbar==4.4.6
django-heroku==0.3.1
google-auth==2.35.0
gunicorn==23.0.0
idna==3.10
jwt==1.3.1
packaging==24.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
python-dotenv==1.0.1
requests>=2.32.3
rsa==4.9
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
dj-database-url
urllib3>=1.25.4,<1.27
whitenoise==6.7.0
boto3>=1.35.44
django-storages==1.14.4